+++ client/karma.conf.js
@@ -45,7 +45,7 @@
                 'src/main/js/util.js',
                 'src/main/js/lib/FileReader.js',
                 'src/main/js/config/config.js',
-                'src/main/js/config/StateDecorator.js',
+                'src/main/js/decorator/StateDecorator.js',
                 'src/main/js/Constants.js',
                 'src/main/js/constants/constantesReserva.js',
                 'src/main/js/constants/constantesLocal.js',
@@ -62,7 +62,6 @@
                 'src/main/js/service/SidebarService.js',
                 'src/main/js/directive/MainToolbarDirective.js',
                 'src/main/js/directive/SidebarDirective.js',
-                'src/main/js/directive/MainFooterDirective.js',
                 'src/main/js/directive/FuncionamentoLocalValidator.js',
                 'src/main/js/controller/AppController.js',
                 'src/main/js/controller/LoginController.js',
@@ -105,7 +104,7 @@
 
             preprocessors: {
                 // Arquivos que serão analisados pela cobertura de testes.
-                'src/main/js/**/!(*lib)/*.js': ['coverage'],
+                'src/main/js/**/!(*lib|*config)/*.js': ['coverage'],
                 'src/main/view/**/*.html': 'ng-html2js'
             },