+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Manutenção: Adicionada licença sobre código.
 * Testes: Criados testes para Toolbar.
 * Refatoração: Refatorado acesso à barra lateral e testes.
+* Testes: Testes e mudanças em Decorator de $state.
 
 ## 0.2.2
 * Issue 82: Corrigida validação da mostragem do botão do menu de Local.